# Modernize faster aws profile print