--- a/CMakeLists.txt
+++ b/CMakeLists.txt
@@ -1,8 +1,21 @@
+# Faster register complete binaries for ndt
+
 cmake_minimum_required(VERSION 3.18)
 
-project("nameless-dt-binaries" VERSION 1.0 LANGUAGES CXX)
+project("nameless-dt-binaries"
+    VERSION 1.0
+    LANGUAGES CXX
+    DESCRIPTION "Faster register complete binaries for nameless-deploy-tools."
+    HOMEPAGE_URL "https://github.com/NitorCreations/nameless-deploy-tools"
+)
 
 set(CMAKE_CXX_STANDARD 20)
+
+# Export compile commands for clang-tidy
+set(CMAKE_EXPORT_COMPILE_COMMANDS ON)
+
+message(STATUS "CMake version: ${CMAKE_VERSION}")
+message(STATUS "CMake source dir: ${CMAKE_SOURCE_DIR}")
 
 include_directories(n_utils)