--- a/n_utils/nameless-dt-print-aws-profiles.cpp
+++ b/n_utils/nameless-dt-print-aws-profiles.cpp
@@ -1,51 +1,49 @@
+#include <filesystem>
 #include <fstream>
 #include <iostream>
 #include <string>
-#include <unordered_set>
+#include <set>
+#include <string_view>
 
-int main(int argc, const char* argv[]) {
-    std::string prefix = "";
-    std::unordered_set<std::string> ret;
+/// Print all unique AWS profile names found from config.
+int main(const int argc, const char* argv[]) {
+    const auto home = std::getenv("HOME");
+    if (!home) {
+        return 0;
+    }
+
+    std::string_view prefix;
+    std::set<std::string> profiles;
     if (argc > 1) {
         prefix = argv[1];
     }
-    if (!getenv("HOME")) {
-        return 0;
-    }
-    std::string home = getenv("HOME");
-    std::ifstream credentials(home + "/.aws/credentials", std::fstream::in);
-    if (credentials) {
-        std::string line;
-        while (getline(credentials, line)) {
-            if (line[0] != '[' || line[line.length() - 1] != ']') {
-                continue;
-            }
-            std::string profile_name = line.substr(1, line.length() - 2);
-            if (strncmp(profile_name.c_str(), prefix.c_str(), prefix.size()) == 0) {
-                ret.insert(profile_name);
-            }
-        }
-        credentials.close();
-    }
-    std::ifstream config(home + "/.aws/config", std::fstream::in);
-    if (config) {
-        std::string line;
-        while (getline(config, line)) {
-            if (line[0] != '[' || line[line.length() - 1] != ']') {
-                continue;
-            }
-            std::string profile_name = line.substr(9, line.length() - 10);
-            if (strncmp(profile_name.c_str(), prefix.c_str(), prefix.size()) == 0) {
-                ret.insert(profile_name);
+
+    // Lambda function to extract aws profile names from a file
+    auto get_aws_profiles = [&](const std::filesystem::path& path, const size_t profile_name_start_offset, const size_t profile_name_end_offset) {
+        std::ifstream file(path, std::fstream::in);
+        if (file) {
+            std::string line;
+            while (std::getline(file, line)) {
+                if (!line.starts_with('[') || !line.ends_with(']')) {
+                    continue;
+                }
+                auto profile_name = std::string_view(line).substr(profile_name_start_offset, line.length() - profile_name_end_offset);
+                if (!profile_name.empty() && profile_name.starts_with(prefix)) {
+                    profiles.insert(std::string(profile_name));
+                }
             }
         }
-        config.close();
+    };
+
+    const std::filesystem::path credentials_path = std::filesystem::path(home) / ".aws" / "credentials";
+    get_aws_profiles(credentials_path, 1, 2);
+
+    const std::filesystem::path config_path = std::filesystem::path(home) / ".aws" / "config";
+    get_aws_profiles(config_path, 9, 10);
+
+    for (const auto& profile : profiles) {
+        std::cout << profile << ' ';
     }
-    std::string res;
-    for (auto profile : ret) {
-        res.append(profile);
-        res.push_back(' ');
-    }
-    std::cout << res << std::endl;
+    std::cout << std::endl;
     return 0;
 }