--- a/n_utils/ndt.py
+++ b/n_utils/ndt.py
@@ -94,7 +94,8 @@ def do_command_completion():
 
 
 def ndt():
-    """The main nameless deploy tools command that provides bash command
+    """
+    The main nameless deploy tools command that provides bash command
     completion and subcommand execution
     """
     do_profile = False