Modernize `nameless-dt-print-aws-profiles.cpp`: use C++17 and C++20 stuff 🤓 

Swapped to the basic `set` from `unordered_set` since set keeps things sorted so nicer print. 

```console
➜  nameless-deploy-tools git:(modernize-profile-cpp-print)
time ndt print-aws-profiles
nitor-core nitor-infra nitor-web nitor-ironbank
ndt print-aws-profiles  0.17s user 0.06s system 48% cpu 0.469 total
➜  nameless-deploy-tools git:(modernize-profile-cpp-print)
time nameless-dt-print-aws-profiles
nitor-core nitor-infra nitor-ironbank nitor-web
nameless-dt-print-aws-profiles  0.00s user 0.00s system 1% cpu 0.219 total
```